Separating helper functions in profile_processing.py into another file and also created a README file, though not completed yet.
